bugs fixed: 397, 496, 499, 500, 502, 503, 505, 506, 508, 509.
git-svn-id: http://co19.googlecode.com/svn/trunk/co19/tests/co19@556 2ecb45c1-dfb3-dc95-330a-13f216d23a04

+++ src/LibTest/core/int/operator_division_A01_t01.dart
@@ -20,7 +20,7 @@ import "dart:math" as Math;
 // @3rdparty sputnik-v1/lib/math_precision.js
 double getPrecision(num number) {
   var log2num = Math.log(number.abs()) / Math.LN2;
-  double pernum = log2num.ceil();
+  int pernum = log2num.ceil();
   return(Math.pow(2, -51 + pernum));
 }
